migration/core/requirements.txt
# Core Dependencies
numpy==1.23.5
pandas==2.0.3
openpyxl==3.1.0
xlsxwriter==3.1.2

# Database
asyncpg==0.28.0

# Async Components
aiohttp==3.8.1
aiofiles==23.2.1

# System & Scheduling
psutil==5.9.0
APScheduler==3.10.1
pytz==2023.3

# Configuration
python-dotenv==0.19.2
pydantic==1.9.0

# Monitoring
prometheus-client==0.16.0

# Testing
pytest==7.1.1
pytest-asyncio==0.18.3
pytest-cov==3.0.0

# FastAPI Stack
fastapi==0.95.0
uvicorn==0.21.1
starlette==0.26.1
python-multipart==0.0.6

# API Documentation
openapi-spec-validator==0.5.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Session Management
itsdangerous==2.1.2